refactor useGenres hook
